Move agnostic funcs from sdkv2 util package
Duplicate CheckArtifactoryLicense func in fw.go since it returns different diag type.
